Zenior parts lisätty automaatio

+++ zenior_saldo_nollassa.php
@@ -36,9 +36,10 @@ if ($yhtiorow["epakurantoinnin_myyntihintaleikkuri"] != 'Z') {
 $query  = "SELECT t.tuoteno,
            a.selitetark        as orig_myyntihinta,
            MAX(t.myyntihinta)  as varahinta,
-           SUM(p.saldo)        as saldosumma
+           SUM(p.saldo)        as saldosumma,
+           t.osasto
            FROM tuote t
-                 JOIN tuotteen_avainsanat a ON (t.yhtio = a.yhtio AND a.kieli = '{$yhtiorow['kieli']}' AND t.tuoteno = a.tuoteno AND a.laji = 'zeniorparts')
+                 JOIN tuotteen_avainsanat a ON (t.yhtio = a.yhtio AND a.kieli = '{$yhtiorow['kieli']}' AND t.tuoteno = a.tuoteno AND a.laji = 'zeniorparts' and a.selite = 'K')
            JOIN tuotepaikat p ON (a.yhtio = p.yhtio AND a.tuoteno = p.tuoteno)
            WHERE t.yhtio           = '{$kukarow["yhtio"]}'
            and t.epakurantti25pvm != '0000-00-00'
@@ -48,12 +49,30 @@ $result = pupe_query($query);
 
 while ($row = mysql_fetch_assoc($result)) {
 
+  if ($yhtiorow['epakurantti_automaattimuutokset'] == 'Z') {
+    $puun_tunnus = t_avainsana("ZEPA_OS_PUUNALK", "", "AND selitetark = (SELECT selitetark FROM tuotteen_avainsanat WHERE yhtio = '{$kukarow["yhtio"]}' AND kieli = '{$yhtiorow["kieli"]}' AND tuoteno = '{$row["tuoteno"]}' AND laji = 'zeniorparts' AND selite = 'ALKUP_OSASTO' LIMIT 1)", "", "", "selite");
+    if ($puun_tunnus != "") {
+      $query = "DELETE FROM puun_alkio WHERE yhtio = '{$kukarow["yhtio"]}' AND liitos = '{$row["tuoteno"]}' AND puun_tunnus = {$puun_tunnus}";
+      pupe_query($query);
+    }
+
+    $set_lisakentat = "try      = (SELECT selitetark FROM tuotteen_avainsanat WHERE yhtio = '{$kukarow["yhtio"]}' AND kieli = '{$yhtiorow["kieli"]}' AND tuoteno = '{$row["tuoteno"]}' AND laji = 'zeniorparts' AND selite = 'ALKUP_TRY' LIMIT 1),
+                       aleryhma = (SELECT selitetark FROM tuotteen_avainsanat WHERE yhtio = '{$kukarow["yhtio"]}' AND kieli = '{$yhtiorow["kieli"]}' AND tuoteno = '{$row["tuoteno"]}' AND laji = 'zeniorparts' AND selite = 'ALKUP_ALERYHMA' LIMIT 1),
+                       osasto   = (SELECT selitetark FROM tuotteen_avainsanat WHERE yhtio = '{$kukarow["yhtio"]}' AND kieli = '{$yhtiorow["kieli"]}' AND tuoteno = '{$row["tuoteno"]}' AND laji = 'zeniorparts' AND selite = 'ALKUP_OSASTO' LIMIT 1),
+                       hinnastoon  = 'E',
+                       ostoehdotus = 'E',";
+  }
+  else {
+    $set_lisakentat = "";
+  }
+
   // jos talteenotettu hinta ei ole nollaa isompi, otetaan viimeisin myyntihinta
   $hinta = (floatval($row["orig_myyntihinta"]) > 0) ? floatval($row["orig_myyntihinta"]) : $row["varahinta"];
   $selite = t("Ep‰kuranttimuutos") . ": ".t("Tuote")." {$row["tuoteno"]} ".t("p‰ivitet‰‰n kurantiksi");
 
   $t_query = "UPDATE tuote
               SET status        = 'T',
+              {$set_lisakentat}
               epakurantti25pvm  = '0000-00-00',
               epakurantti50pvm  = '0000-00-00',
               epakurantti75pvm  = '0000-00-00',
@@ -63,12 +82,14 @@ while ($row = mysql_fetch_assoc($result)) {
               AND tuoteno       = '{$row["tuoteno"]}';";
   $t_result = pupe_query($t_query);
 
-  $t_query = "DELETE FROM tuotteen_avainsanat
-              WHERE yhtio = '{$kukarow["yhtio"]}'
-              AND kieli   = '{$yhtiorow["kieli"]}'
-              AND laji    = 'zeniorparts'
-              AND tuoteno = '{$row["tuoteno"]}';";
-  $t_result = pupe_query($t_query);
+  if ($yhtiorow['epakurantti_automaattimuutokset'] == 'Z') {
+    $t_query = "DELETE FROM tuotteen_avainsanat
+                WHERE yhtio   = '{$kukarow["yhtio"]}'
+                  AND kieli   = '{$yhtiorow["kieli"]}'
+                  AND laji    = 'zeniorparts'
+                  AND tuoteno = '{$row["tuoteno"]}';";
+    $t_result = pupe_query($t_query);
+  }
 
   $t_query = "INSERT INTO tapahtuma SET
               yhtio    = '{$kukarow["yhtio"]}',
